# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hospital SaaS API",
    "description": "Public Swagger documentation for Hospital SaaS Backend",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development"
    },
    {
      "url": "https://YOUR_PRODUCTION_DOMAIN",
      "description": "Production (Replace manually)"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Login successful" } }
      }
    },
    "/api/auth/logout": {
      "post": {
        "security": [{ "BearerAuth": [] }],
        "summary": "Logout",
        "responses": { "200": { "description": "Logged out" } }
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "summary": "Forgot Password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Reset email sent" } }
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "summary": "Reset Password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" },
                  "newPassword": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Password reset" } }
      }
    },
...
  }
}